+++ lib/puppet/type/cf_puppetserver.rb
@@ -1,0 +1,50 @@
+Puppet::Type.newtype(:cf_puppetserver) do
+    desc "DO NOT USE DIRECTLY."
+    
+    autorequire(:cfsystem_flush_config) do
+        ['begin']
+    end
+    autorequire(:cfsystem_memory_calc) do
+        ['total']
+    end
+    autonotify(:cfsystem_flush_config) do
+        ['commit']
+    end
+    
+    ensurable do
+        defaultvalues
+        defaultto :absent
+    end
+    
+    newparam(:name) do
+        isnamevar
+    end
+    
+    newproperty(:service_name) do
+        isrequired
+        validate do |value|
+            unless value =~ /^[a-z0-9_@-]+$/i
+                raise ArgumentError, "%s is not a valid service name" % value
+            end
+        end
+    end
+
+    newproperty(:cpu_weight) do
+        isrequired
+        validate do |value|
+            unless value.is_a? Integer and value > 0
+                raise ArgumentError, "%s is not a valid positive integer" % value
+            end
+        end
+    end
+    
+    newproperty(:io_weight) do
+        isrequired
+        validate do |value|
+            unless value.is_a? Integer and value > 0
+                raise ArgumentError, "%s is not a valid positive integer" % value
+            end
+        end
+    end
+    
+end